SegCurrent2Field.cpp
CUDA function to calculate a 3D magnetic field distribution. A segmented version of Current2Field, meaning that this function is used when the size of the space volume is bigger than the GPU grid size. The volume is divided into smaller regions to accommodate the calculation of the GPU specifications.